Extend the tail of a distribution using Generalized Pareto Distribution (GPD)

This function takes existing observation vectors and extends them by generating extra values based on GPD parameters. It combines the original data with newly generated tail data.

c_extendtail(uni, w, mlesi, mlesh, extra_n)

Arguments

uni

A NumericVector of existing observation values.

w

A NumericVector of existing probabilities/weights.

mlesi

Double; the scale parameter for the GPD.

mlesh

Double; the shape parameter for the GPD.

extra_n

Double; the number of extra points to generate for the tail.

Value

A named list containing:

uni: A combined numeric vector of original and extended values.

w: A combined numeric vector of associated probabilities/weights.

Examples

if (FALSE) { # \dontrun{
data(ipilimumab)
dat <- ipilimumab

### Read the data
X <- dat %>% filter(arm=="ipilimumab") %>% pull(time)
Xs <- dat %>% filter(arm=="ipilimumab") %>% pull(event)

### Remove the last few events to illustrate the functionality
wher <- X>26
Xs[wher] <- 0

### Estimate and plot the Kaplan-Meier estimation of the ECDF
fit <- c_ecdf_surv(X,Xs)
plot(w~uni,data=fit,type="s",ylim=c(0,1))
abline(h=1,lty=3)

### Estimate the scale and shape parameters for the GPD
p <- param_gpd(X,Xs)

### Extend the tail
uni <- fit$uni
w <- fit$w
mlesi <- p$mlesi
mlesh <- p$mlesh
ta <- c_extendtail(uni,w,mlesi,mlesh,p$extra_n)
ta

### plot the extension
plot(w~uni,data=ta,type="s",col="red",ylim=c(0,1),xlim = c(0,max(ta$uni)))
points(w~uni,data=fit,type="s")
abline(h=1,lty=3)

###############
### Example 2
###############
### Read the data
X <- dat %>% filter(arm=="placebo") %>% pull(time)
Xs <- dat %>% filter(arm=="placebo") %>% pull(event)

### Remove the last few events to illustrate the functionality
wher <- X>10
Xs[wher] <- 0

### Estimate aplan-Meier estimation of the ECDF
fit <- c_ecdf_surv(X,Xs)

### Estimate the scale and shape parameters for the GPD
p <- param_gpd(X,Xs)

### Extend the tail
ta <- c_extendtail(fit$uni,fit$w,p$mlesi,p$mlesh,extra_n=10)
ta

### plot the extension
plot(w~uni,data=ta,type="s",col="red",ylim=c(0,1),xlim = c(0,max(ta$uni)))
points(w~uni,data=fit,type="s")
abline(h=1,lty=3)

### Extend the tail but only with 5 values
ta <- c_extendtail(fit$uni,fit$w,p$mlesi,p$mlesh,extra_n=5)
ta

### plot the extension
plot(w~uni,data=ta,type="s",col="red",ylim=c(0,1),xlim = c(0,max(ta$uni)))
points(w~uni,data=fit,type="s")
abline(h=1,lty=3)
} # }